Create script_run_notebook.ipynb
script_run_notebook.ipynb
The script_run_notebook.ipynb notebook serves as an easy way to execute script.py on an Azure ML compute. This notebook is similar to the existing automated ML sample notebooks however, there are a couple of key differences as explained in the following sections.

Environment
Typically, the training environment for an automated ML run is automatically set by the SDK. However, when running a custom script run like the generated code, automated ML is no longer driving the process, so the environment must be specified for the script run to succeed.

Code generation reuses the environment that was used in the original automated ML experiment, if possible. Doing so guarantees that the training script run doesn't fail due to missing dependencies, and has a side benefit of not needing a Docker image rebuild, which saves time and compute resources.

If you make changes to script.py that require additional dependencies, or you would like to use your own environment, you need to update the Create environment cell in script_run_notebook.ipynb accordingly.

For more information about AzureML environments, see the Environment class documentation.

Submit the experiment
Since the generated code isn’t driven by automated ML anymore, instead of creating an AutoMLConfig and then passing it to experiment.submit(), you need to create a ScriptRunConfig and provide the generated code (script.py) to it.

The following example contains the parameters and regular dependencies needed to run ScriptRunConfig, such as compute, environment, etc. For more information on how to use ScriptRunConfig, see Configure and submit training runs.